Add backup task folder and modify existing task generation functionality
Previously, the code added a backup task folder to the master object's attributes. The code also modified the get_existed_task method to include the backup task folder in the folder list. The exist_task method was modified to check if a task with a given name exists. The get_exist_task_path method was modified to return a list of paths of existing tasks with a given JSON name. The generate_tasks and generate_manual_tasks methods were modified to use the new functionality of checking if a task with a given name exists before generating it.